Remove single table inheritance LM design
Learning materials are treated as a single entity for clarity and ease
of use.
